basic page layouts, relatively broken haha
