--- .env.local.example
@@ -1,5 +1,0 @@
-CONTENTFUL_SPACE_ID=
-CONTENTFUL_ACCESS_TOKEN=
-CONTENTFUL_PREVIEW_ACCESS_TOKEN=
-CONTENTFUL_PREVIEW_SECRET=
-CONTENTFUL_REVALIDATE_SECRET=
