+++ components/alert.js
@@ -1,40 +1,11 @@
 import Container from './container';
 import cn from 'classnames';
 
-export default function Alert({ preview }) {
+export default function Alert({ message }) {
   return (
-    <div
-      className={cn('border-b', {
-        'bg-accent-7 border-accent-7 text-white': preview,
-        'bg-accent-1 border-accent-2': !preview,
-      })}
-    >
+    <div className={cn('border-b bg-accent-1 border-accent-2')}>
       <Container>
-        <div className='py-2 text-center text-sm'>
-          {preview ? (
-            <>
-              This is page is a preview.{' '}
-              <a
-                href='/api/exit-preview'
-                className='underline hover:text-cyan duration-200 transition-colors'
-              >
-                Click here
-              </a>{' '}
-              to exit preview mode.
-            </>
-          ) : (
-            <>
-              The source code for this tool is{' '}
-              <a
-                href={`https://https://github.com/qwoodmansee/bingot-cms`}
-                className='underline hover:text-success duration-200 transition-colors'
-              >
-                available on GitHub
-              </a>
-              .
-            </>
-          )}
-        </div>
+        <div className='py-2 text-center text-sm'>{message}</div>
       </Container>
     </div>
   );
