+++ components/layout.js
@@ -1,4 +1,4 @@
-import Alert from '../components/alert';
+import { PreviewModeAlert } from '../components/preview-mode-alert';
 import Footer from '../components/footer';
 import Meta from '../components/meta';
 import { BingoTeacherNavBar } from './bingo-teacher-components/bingoTeacherNavBar';
@@ -7,9 +7,9 @@ export default function Layout({ preview, children }) {
   return (
     <>
       <Meta />
+      <BingoTeacherNavBar />
       <div className='min-h-screen'>
-        <BingoTeacherNavBar />
-        <Alert preview={preview} />
+        {preview && <PreviewModeAlert />}
         <main>{children}</main>
       </div>
       <Footer />
