+++ package.json
@@ -24,6 +24,7 @@
     "autoprefixer": "10.4.2",
     "contentful-import": "^7.8.7",
     "eslint": "^8.33.0",
+    "eslint-config-next": "^13.1.6",
     "postcss": "8.4.5",
     "tailwindcss": "^3.2.4",
     "typescript": "^4.9.5"
