+++ pages/blog/index.tsx
@@ -1,0 +1,52 @@
+import { useRouter } from 'next/router';
+import Head from 'next/head';
+import ErrorPage from 'next/error';
+import Container from '../../components/container';
+import PostBody from '../../components/post-body';
+import MoreStories from '../../components/more-stories';
+import Header from '../../components/header';
+import PostHeader from '../../components/post-header';
+import SectionSeparator from '../../components/section-separator';
+import Layout from '../../components/layout';
+import {
+  getAllPostsForHome,
+  getAllPostsWithSlug,
+  getPostAndMorePosts,
+} from '../../lib/api';
+import PostTitle from '../../components/post-title';
+import HeroPost from '../../components/hero-post';
+import Intro from '../../components/intro';
+
+export default function BlogHome({ allPosts, preview }) {
+  const heroPost = allPosts[0];
+  const morePosts = allPosts.slice(1);
+
+  return (
+    <>
+      <Head>
+        <title>BingoTeacher</title>
+      </Head>
+      <Container>
+        <Intro />
+        {heroPost && (
+          <HeroPost
+            title={heroPost.title}
+            coverImage={heroPost.coverImage}
+            date={heroPost.date}
+            author={heroPost.author}
+            slug={heroPost.slug}
+            excerpt={heroPost.excerpt}
+          />
+        )}
+        {morePosts.length > 0 && <MoreStories posts={morePosts} />}
+      </Container>
+    </>
+  );
+}
+
+export async function getStaticProps({ preview = false }) {
+  const allPosts = (await getAllPostsForHome(preview)) ?? [];
+  return {
+    props: { preview, allPosts },
+  };
+}
