+++ pages/blog/posts/[slug].tsx
@@ -1,15 +1,15 @@
 import { useRouter } from 'next/router';
 import Head from 'next/head';
 import ErrorPage from 'next/error';
-import Container from '../../components/container';
-import PostBody from '../../components/post-body';
-import MoreStories from '../../components/more-stories';
-import Header from '../../components/header';
-import PostHeader from '../../components/post-header';
-import SectionSeparator from '../../components/section-separator';
-import Layout from '../../components/layout';
-import { getAllPostsWithSlug, getPostAndMorePosts } from '../../lib/api';
-import PostTitle from '../../components/post-title';
+import Container from '../../../components/container';
+import PostBody from '../../../components/post-body';
+import MoreStories from '../../../components/more-stories';
+import Header from '../../../components/header';
+import PostHeader from '../../../components/post-header';
+import SectionSeparator from '../../../components/section-separator';
+import Layout from '../../../components/layout';
+import { getAllPostsWithSlug, getPostAndMorePosts } from '../../../lib/api';
+import PostTitle from '../../../components/post-title';
 export default function Post({ post, morePosts, preview }) {
   const router = useRouter();
