+++ pages/index.tsx
@@ -6,37 +6,10 @@ import Layout from '../components/layout';
 import { getAllPostsForHome } from '../lib/api';
 import Head from 'next/head';
 
-export default function Index({ preview, allPosts }) {
-  const heroPost = allPosts[0];
-  const morePosts = allPosts.slice(1);
+export default function Index({ preview, children }) {
   return (
     <>
-      <Layout preview={preview}>
-        <Head>
-          <title>BingoTeacher</title>
-        </Head>
-        <Container>
-          <Intro />
-          {heroPost && (
-            <HeroPost
-              title={heroPost.title}
-              coverImage={heroPost.coverImage}
-              date={heroPost.date}
-              author={heroPost.author}
-              slug={heroPost.slug}
-              excerpt={heroPost.excerpt}
-            />
-          )}
-          {morePosts.length > 0 && <MoreStories posts={morePosts} />}
-        </Container>
-      </Layout>
+      <Layout preview={preview}>{children}</Layout>
     </>
   );
-}
-
-export async function getStaticProps({ preview = false }) {
-  const allPosts = (await getAllPostsForHome(preview)) ?? [];
-  return {
-    props: { preview, allPosts },
-  };
 }
