+++ pages/user/login/index.tsx
@@ -1,0 +1,3 @@
+export default function Login() {
+  return <h1>Login Screen</h1>;
+}
